learning_footpaths_backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
python-dotenv==1.0.1
redis==5.2.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.6
Flask-Migrate
psycopg2-binary